back-end/requirements.txt
Django==3.1.7
djangorestframework==3.12.2
django-cors-headers==3.7.0
django-filter==2.4.0
django-mass-edit==3.2.0
django-import-export==2.5.0
drf-firebase-auth==1.0.0
psycopg2==2.8.6
